Rename package of shardingsphere-infra-cluster-mode-repository-provider
