+++ shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-cluster-mode/shardingsphere-infra-cluster-mode-repository/shardingsphere-infra-cluster-mode-repository-provider/shardingsphere-infra-cluster-mode-repository-zookeeper-curator/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.mode.repository.cluster.ClusterPersistRepository
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.governance.repository.zookeeper.CuratorZookeeperRepository
+org.apache.shardingsphere.infra.mode.repository.cluster.zookeeper.CuratorZookeeperRepository
