--- shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-cluster-mode/shardingsphere-infra-cluster-mode-repository/shardingsphere-infra-cluster-mode-repository-provider/shardingsphere-infra-cluster-mode-repository-zookeeper-curator/src/test/java/org/apache/shardingsphere/governance/repository/zookeeper/handler/CuratorZookeeperExceptionHandlerTest.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.governance.repository.zookeeper.handler;
-
-import org.apache.shardingsphere.infra.mode.repository.cluster.ClusterPersistRepositoryException;
-import org.apache.zookeeper.KeeperException.NoNodeException;
-import org.junit.Test;
-
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
-
-public final class CuratorZookeeperExceptionHandlerTest {
-    
-    @Test
-    public void assertHandleException() {
-        CuratorZookeeperExceptionHandler.handleException(null);
-        CuratorZookeeperExceptionHandler.handleException(new NoNodeException(""));
-        CuratorZookeeperExceptionHandler.handleException(new Exception(new NoNodeException("")));
-        try {
-            CuratorZookeeperExceptionHandler.handleException(new Exception(""));
-            fail("must be failed handle new Exception().");
-            // CHECKSTYLE:OFF
-        } catch (final Exception ex) {
-            // CHECKSTYLE:ON
-            assertTrue(ex instanceof ClusterPersistRepositoryException);
-        }
-    }
-}
